\name{gene2}
\alias{gene2}
\docType{data}
\title{gene2 dataset}
\description{A list of genotype, quantitative phenotypes, binary phenotypes, and causal variants. The gene2 was described in the reference below. The phenotypes were generated by simulation models with known causal variants.}
\usage{data(gene2)}
\format{
A list of 4 variables:

 geno: a genotype matrix of 1969 by 98
 
 phen: a matrix of 1969 by 5, each column is the simulated quantitative phenotypes in different scenario.

 phen.bi: a matrix of 1969 by 5, each column is the simulated binary phenotypes in different scenario.

 causal: information about the causal variants used in the simulated phenotypes.
}
\references{
C. Xu, M. Ladouceur, Z. Dastani, J. B. Richards, A. Ciampi, C. M.T. Greenwood (2012), Multiple regression methods show great potential for rare variant association tests, PLoSONE.
}
\examples{
data(gene2)
str(gene2) 
}
\keyword{datasets}
