\name{gene3}
\alias{gene3}
\docType{data}
\title{gene3 dataset}
\description{A list of genotype, quantitative phenotypes, binary phenotypes, and causal variants. The gene3 was described in the reference below. The phenotypes were generated by simulation models with known causal variants.}
\usage{data(gene3)}
\format{
A list of 4 variables:

 geno: a genotype matrix of 1981 by 28
 
 phen: a matrix of 1981 by 5, each column is the simulated quantitative phenotypes in different scenario.

 phen.bi: a matrix of 1981 by 5, each column is the simulated binary phenotypes in different scenario.

 causal: information about the causal variants used in the simulated phenotypes. 
}
\references{
C. Xu, M. Ladouceur, Z. Dastani, J. B. Richards, A. Ciampi, C. M.T. Greenwood (2012), Multiple regression methods show great potential for rare variant association tests, PLoSONE.
}
\examples{
data(gene3)
str(gene3) 
}
\keyword{datasets}
