\name{geno2count}
\alias{geno2count}
\title{Transforming genotype codes matrix to genotype counts}
\description{Genotype counts }
\usage{
geno2count(genotype)
}
\arguments{
  \item{genotype}{ Matrix}
}
\value{
Data frame of genotype counts with 3 columns: indid, snpid, and count
}
\author{C. Xu}
\seealso{
\code{\link{count2geno}}
}
\keyword{datasets}